+++ cjapy/cjapy.py
@@ -1490,7 +1490,7 @@ class CJA:
         if full:
             params[
                 "expansion"
-            ] = "shares,tags,accessLevel,modified,externalReferences,definition"
+            ] = "shares,tags,accessLevel,modified,externalReferences,definition,ownerFullName,sharesFullName,complexity,usageSummary"
         if filterByIds:
             params["filterByIds"] = filterByIds
         if ownerId: